:heavy_plus_sign: Add Google Formatter, updates in README, as well as minor changes in other code
