+++ src/main/java/frc/team4077/robot/common/NavX.java
@@ -3,80 +3,81 @@ package frc.team4077.robot.common;
 import com.kauailabs.navx.AHRSProtocol.AHRSUpdateBase;
 import com.kauailabs.navx.frc.AHRS;
 import com.kauailabs.navx.frc.ITimestampedDataSubscriber;
-
 import edu.wpi.first.wpilibj.SPI;
 
-/**
- * Driver for a NavX board. Basically a wrapper for the {@link AHRS} class
- */
+/** Driver for a NavX board. Basically a wrapper for the {@link AHRS} class */
 public class NavX {
-	protected class Callback implements ITimestampedDataSubscriber {
-		@Override
-		public void timestampedDataReceived(long system_timestamp, long sensor_timestamp, AHRSUpdateBase update,
-				Object context) {
-			synchronized (NavX.this) {
-				// This handles the fact that the sensor is inverted from our coordinate conventions.
-				if (mLastSensorTimestampMs != kInvalidTimestamp && mLastSensorTimestampMs < sensor_timestamp) {
-					mYawRateDegreesPerSecond = 1000.0 * (-mYawDegrees - update.yaw)
-							/ (double) (sensor_timestamp - mLastSensorTimestampMs);
-				}
-				mLastSensorTimestampMs = sensor_timestamp;
-				mYawDegrees = -update.yaw;
-			}
-		}
-	}
+  protected class Callback implements ITimestampedDataSubscriber {
+    @Override
+    public void timestampedDataReceived(
+        long system_timestamp, long sensor_timestamp, AHRSUpdateBase update, Object context) {
+      synchronized (NavX.this) {
+        // This handles the fact that the sensor is inverted from our coordinate
+        // conventions.
+        if (mLastSensorTimestampMs != kInvalidTimestamp
+            && mLastSensorTimestampMs < sensor_timestamp) {
+          mYawRateDegreesPerSecond =
+              1000.0
+                  * (-mYawDegrees - update.yaw)
+                  / (double) (sensor_timestamp - mLastSensorTimestampMs);
+        }
+        mLastSensorTimestampMs = sensor_timestamp;
+        mYawDegrees = -update.yaw;
+      }
+    }
+  }
 
-	protected AHRS mAHRS;
+  protected AHRS mAHRS;
 
-	protected double mAngleAdjustment;
-	protected double mYawDegrees;
-	protected double mYawRateDegreesPerSecond;
-	protected final long kInvalidTimestamp = -1;
-	protected long mLastSensorTimestampMs;
+  protected double mAngleAdjustment;
+  protected double mYawDegrees;
+  protected double mYawRateDegreesPerSecond;
+  protected final long kInvalidTimestamp = -1;
+  protected long mLastSensorTimestampMs;
 
-	public NavX(SPI.Port spi_port_id) {
-		mAHRS = new AHRS(spi_port_id, (byte) 200);
-		resetState();
-		mAHRS.registerCallback(new Callback(), null);
-	}
+  public NavX(SPI.Port spi_port_id) {
+    mAHRS = new AHRS(spi_port_id, (byte) 200);
+    resetState();
+    mAHRS.registerCallback(new Callback(), null);
+  }
 
-	public synchronized void reset() {
-		mAHRS.reset();
-		resetState();
-	}
+  public synchronized void reset() {
+    mAHRS.reset();
+    resetState();
+  }
 
-	public synchronized void zeroYaw() {
-		mAHRS.zeroYaw();
-		resetState();
-	}
+  public synchronized void zeroYaw() {
+    mAHRS.zeroYaw();
+    resetState();
+  }
 
-	private void resetState() {
-		mLastSensorTimestampMs = kInvalidTimestamp;
-		mYawDegrees = 0.0;
-		mYawRateDegreesPerSecond = 0.0;
-	}
+  private void resetState() {
+    mLastSensorTimestampMs = kInvalidTimestamp;
+    mYawDegrees = 0.0;
+    mYawRateDegreesPerSecond = 0.0;
+  }
 
-	public synchronized void setAngleAdjustment(double adjustment) {
-		mAngleAdjustment = adjustment;
-	}
+  public synchronized void setAngleAdjustment(double adjustment) {
+    mAngleAdjustment = adjustment;
+  }
 
-	protected synchronized double getRawYawDegrees() {
-		return mYawDegrees;
-	}
+  protected synchronized double getRawYawDegrees() {
+    return mYawDegrees;
+  }
 
-	public double getYaw() {
-		return mAngleAdjustment + getRawYawDegrees();
-	}
+  public double getYaw() {
+    return mAngleAdjustment + getRawYawDegrees();
+  }
 
-	public double getYawRateDegreesPerSec() {
-		return mYawRateDegreesPerSecond;
-	}
+  public double getYawRateDegreesPerSec() {
+    return mYawRateDegreesPerSecond;
+  }
 
-	public double getYawRateRadiansPerSec() {
-		return 180.0 / Math.PI * getYawRateDegreesPerSec();
-	}
+  public double getYawRateRadiansPerSec() {
+    return 180.0 / Math.PI * getYawRateDegreesPerSec();
+  }
 
-	public double getRawAccelX() {
-		return mAHRS.getRawAccelX();
-	}
+  public double getRawAccelX() {
+    return mAHRS.getRawAccelX();
+  }
 }
