+++ src/main/java/frc/team4077/robot/common/RobotMap.java
@@ -6,9 +6,9 @@ package frc.team4077.robot.common;
  * @author FRC 4077 MASH, Hillel Coates
  */
 public class RobotMap {
-	// Drive
-	public static final int LEFT_MASTER_TALON_ID = 1;
-	public static final int RIGHT_MASTER_TALON_ID = 3;
-	public static final int LEFT_SLAVE_TALON_ID = 2;
-	public static final int RIGHT_SLAVE_TALON_ID = 4;
+  // Drive
+  public static final int LEFT_MASTER_TALON_ID = 1;
+  public static final int RIGHT_MASTER_TALON_ID = 3;
+  public static final int LEFT_SLAVE_TALON_ID = 2;
+  public static final int RIGHT_SLAVE_TALON_ID = 4;
 }
